--- a/cmake/Dependencies.cmake
+++ b/cmake/Dependencies.cmake
@@ -18,7 +18,7 @@ FetchContent_Declare(
 FetchContent_Declare(
         fmt
         GIT_REPOSITORY https://github.com/fmtlib/fmt.git
-        GIT_TAG 10.0.0
+        GIT_TAG 10.1.1
 )
 
 # dependency of mp-units, building examples, tests, etc is off by default