## IoBuffer & IoSerialiserYaS improvements & {fmt} format fixes

* bumped fmt version and fixed formatter
* IoBuffer: simplified templating and added endianness awareness, aiming at:
  * minimising use of reinterpret_cast (or at least limit this to fewer cases)
  * minimising use of raw pointer in public API
  * reducing non-functional code-duplication
* IoSerialiserYaS: removed unused fieldname hash


No performance regression, rather a slight improvement on all serialisers:
```cpp
Before:
┌─protocol─┬────────ALWAYS─────────┬────────LENIENT────────┬────────IGNORE─────────┐
│   YAML   │131.8 MB/s ± 547.8 kB/s│127.4 MB/s ± 147.6 kB/s│132.1 MB/s ± 190.4 kB/s│
│   Json   │305.1 MB/s ±   1.6 MB/s│301.4 MB/s ±   1.5 MB/s│305.4 MB/s ±   2.1 MB/s│
│   YaS    │  7.0 GB/s ±  14.6 MB/s│  6.4 GB/s ±   6.0 MB/s│  7.8 GB/s ±  10.2 MB/s│
│ CmwLight │  8.1 GB/s ±  13.7 MB/s│  7.7 GB/s ±  11.9 MB/s│  8.5 GB/s ±  12.1 MB/s│
└──────────┴───────────────────────┴───────────────────────┴───────────────────────┘
After IoBuffer clean-up:
┌─protocol─┬────────ALWAYS─────────┬────────LENIENT────────┬────────IGNORE─────────┐
│   YAML   │127.2 MB/s ±   2.0 MB/s│129.5 MB/s ± 401.7 kB/s│127.5 MB/s ± 272.4 kB/s│
│   Json   │314.3 MB/s ±   1.5 MB/s│315.0 MB/s ±   1.8 MB/s│316.4 MB/s ±   1.8 MB/s│
│   YaS    │  6.9 GB/s ±  36.5 MB/s│  6.4 GB/s ±  18.6 MB/s│  7.5 GB/s ±  40.3 MB/s│
│ CmwLight │  8.0 GB/s ±  40.2 MB/s│  7.7 GB/s ±  54.9 MB/s│  8.5 GB/s ±  39.0 MB/s│
└──────────┴───────────────────────┴───────────────────────┴───────────────────────┘
After YaS unused fieldname hash removal:
┌─protocol─┬────────ALWAYS─────────┬────────LENIENT────────┬────────IGNORE─────────┐
│   YAML   │126.9 MB/s ± 582.4 kB/s│127.0 MB/s ± 305.6 kB/s│127.0 MB/s ± 420.9 kB/s│
│   Json   │300.2 MB/s ±   2.2 MB/s│303.6 MB/s ±   1.8 MB/s│301.5 MB/s ±   1.9 MB/s│
│   YaS    │  7.2 GB/s ±  30.5 MB/s│  6.6 GB/s ±  20.5 MB/s│  8.0 GB/s ±  37.0 MB/s│
│ CmwLight │  8.0 GB/s ±  36.2 MB/s│  7.7 GB/s ±  17.3 MB/s│  8.7 GB/s ±  68.3 MB/s│
└──────────┴───────────────────────┴───────────────────────┴───────────────────────┘
````